simplify: dedupe, parallelise, align names with DraftChanges
Code-reuse and quality cleanup over the harmonization commits, plus
rename to leave the QuickSetup section API a strict superset of
DraftChanges' ConfigSection API.

Dedupe and placement
- `cssHeroCard` exported once from `SetupCard.ts`, dropped from
  `AuthenticationSection.ts` (was byte-identical).
- `cssQuickSetupCard` and `cssShadowedPrimaryButton` moved out of
  `QuickSetupStepHeader.ts` (a "header" file) into `SettingsLayout.ts`
  next to `cssCardSurface` and `cardSurfaceShadow`.
- `SandboxSection._needsRestart()` was a method that built a fresh
  `Computed` per call (latent leak); now a `protected readonly` field.

Efficiency
- `AuthenticationSection.apply()` runs `_fetchProviders` and
  `_fetchPrefsPendingChanges` in parallel post-restart instead of
  sequentially.
- `SandboxSection.apply()` drops its `delay(2000)` wait to `delay(500)`
  before `waitUntilReady`. (Auth and `DraftChanges` use no delay; this
  is a sandbox-only workaround for `waitUntilReady` catching the
  pre-restart server. Worth fixing in `ConfigAPI` next.)

Naming
- `QuickSetupSection.hasPendingChanges` -> `isDirty`.
- `QuickSetupSection.applyPendingChanges()` -> `apply()`.
  Now matches `ConfigSection.isDirty` / `ConfigSection.apply()` in
  `DraftChanges.ts`. `isApplying` was already aligned with
  `DraftChangesManager.isApplying`. The `QuickSetupSection` interface
  now reads as a strict superset of `ConfigSection` (adds `canProceed`,
  optional `customLabel`, public `isApplying`), keeping a future
  unification (each step uses a `DraftChangesManager`) a small diff.
